move 'http.client' to 'httpclient' for consistency
